[Refactor] Extract litegraph canvas drop target logic as composable
